Move ClientInput/ServerOutput into Hydra.API
When referring to these modules/datatypes it occurred to me that putting
them into the `Hydra.API` scope makes it clear they are part of that layer.
