+++ hydra-node/src/Hydra/API/Server.hs
@@ -17,12 +17,12 @@ import Control.Concurrent.STM.TChan (newBroadcastTChanIO, writeTChan)
 import Control.Concurrent.STM.TVar (TVar, modifyTVar', newTVarIO, readTVar)
 import Control.Exception (IOException)
 import qualified Data.Aeson as Aeson
-import Hydra.ClientInput (ClientInput)
+import Hydra.API.ClientInput (ClientInput)
 import Hydra.Ledger (IsTx (..))
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.Network (IP, PortNumber)
 import Hydra.Party (Party)
-import Hydra.ServerOutput (ServerOutput (Greetings, InvalidInput))
+import Hydra.API.ServerOutput (ServerOutput (Greetings, InvalidInput))
 import Network.WebSockets (
   acceptRequest,
   receiveData,
