+++ hydra-node/src/Hydra/Logging/Monitoring.hs
@@ -26,7 +26,7 @@ import Hydra.Logging.Messages (HydraLog (..))
 import Hydra.Network (PortNumber)
 import Hydra.Network.Message (Message (ReqTx))
 import Hydra.Node (HydraNodeLog (BeginEvent, EndEffect, EndEvent))
-import Hydra.ServerOutput (ServerOutput (..))
+import Hydra.API.ServerOutput (ServerOutput (..))
 import Hydra.Snapshot (Snapshot (confirmed))
 import System.Metrics.Prometheus.Http.Scrape (serveMetrics)
 import System.Metrics.Prometheus.Metric (Metric (CounterMetric, HistogramMetric))
