+++ hydraw/src/Hydra/Painter.hs
@@ -7,11 +7,11 @@ import qualified Data.Aeson as Aeson
 import qualified Data.Map as Map
 import Hydra.Cardano.Api
 import Hydra.Chain.Direct.Util (readFileTextEnvelopeThrow)
-import Hydra.ClientInput (ClientInput (GetUTxO, NewTx))
+import Hydra.API.ClientInput (ClientInput (GetUTxO, NewTx))
 import Hydra.Ledger.Cardano (emptyTxBody)
 import Hydra.Network (Host (..))
 import Hydra.Prelude
-import Hydra.ServerOutput (ServerOutput (GetUTxOResponse))
+import Hydra.API.ServerOutput (ServerOutput (GetUTxOResponse))
 import Network.WebSockets (
   Connection,
   runClient,
